Refactor and modularize search, filter, and modal functionalities
The `searchAndFilter`, `sortCharAR`, and utility functions have been moved to dedicated service files for better modularity. Event and component modules have been updated to improve cohesion and reduce dependencies on the main index file.
